Initial commit new recently closed urls implementation.
A list of closed urls are saved on CLOSED_URL_FILENAME file. Upon closing a web page
its URL is appended to this list. When pressing GDK_U (revive) the URL is removed
from the list and a new instance opens up. The list acts as a LIFO structure on disk.

For this to work all vimprobable instances are in charge of determining when to discard
the contents of this list (as it can grow largely); and empty it when there are no more
instances left. The basic implementation is a "session" (RUNNING_INSTANCES_FILENAME)
file which keeps track of the number of instances.